pyproject.toml
[project]
name = "oau-bike-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "aiosmtplib>=4.0.1",
    "alembic>=1.16.1",
    "alembic-postgresql-enum>=1.7.0",
    "asyncpg>=0.30.0",
    "databases[asyncpg]>=0.9.0",
    "fastapi[standard]>=0.115.12",
    "geopy>=2.4.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "shapely>=2.1.1",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.34.3",
    "websockets>=15.0.1",
]
